+++ sample/sample_benchmark_task_malloc.cpp
@@ -77,7 +77,7 @@ int main(int argc, char *argv[]) {
   }
 
   if (argc > 3) {
-    g_stack_size = atoi(argv[3]) * 1024;
+    g_stack_size = static_cast<size_t>(atoi(argv[3]) * 1024);
   }
   if (g_stack_size < copp::stack_traits::minimum_size()) {
     g_stack_size = copp::stack_traits::minimum_size();
@@ -103,7 +103,7 @@ int main(int argc, char *argv[]) {
 
   // start a task
   for (int i = 0; i < max_task_number; ++i) {
-    task_arr[i]->start();
+    task_arr[static_cast<size_t>(i)]->start();
   }
 
   // yield & resume from runner
@@ -113,10 +113,10 @@ int main(int argc, char *argv[]) {
   while (continue_flag) {
     continue_flag = false;
     for (int i = 0; i < max_task_number; ++i) {
-      if (false == task_arr[i]->is_completed()) {
+      if (false == task_arr[static_cast<size_t>(i)]->is_completed()) {
         continue_flag = true;
         ++real_switch_times;
-        task_arr[i]->resume();
+        task_arr[static_cast<size_t>(i)]->resume();
       }
     }
   }
